+++ CHANGELOG.md
@@ -83,6 +83,7 @@ In this phase (matterbridge 3.4.x) all plugins will not build and will not run w
 - [frontend]: Bumped `frontend` version to 3.2.0.
 - [frontend]: General improvements and small bug fixes.
 - [frontend]: Added SystemInfo to Settings.
+- [frontend]: Added RvcRunMode to IconView.
 
 ### Changed